refactor(subject-select): move subject selection to own component (#41)
