+++ app/protected/start/page.tsx
@@ -8,36 +8,12 @@ import {
     CardHeader,
     CardTitle,
 } from '@/components/ui/card';
-import { Subject, subjectIcons } from '../../constants/subjects';
-import SubjectCheckbox from '@/components/subject-checkbox';
+import SubjectSelect from '@/components/ui/subjectSelect';
 import { ChevronRight } from 'lucide-react';
-import { useLocalStorage } from 'react-use';
-import { useState, useEffect } from 'react';
 import Image from 'next/image';
 import Link from 'next/link';
 
 export default function StartPage() {
-    // Store selected subjects in local storage
-    const [selectedSubjects, setSelectedSubjects] = useLocalStorage<Subject[]>(
-        'selectedSubjects',
-        [],
-    );
-
-    // Track hydration to prevent SSR mismatch
-    const [isHydrated, setIsHydrated] = useState(false);
-
-    useEffect(() => {
-        setIsHydrated(true);
-    }, []);
-
-    function toggleSubject(subject: Subject) {
-        if (selectedSubjects?.includes(subject)) {
-            setSelectedSubjects(selectedSubjects.filter((s) => s !== subject));
-        } else {
-            setSelectedSubjects([...(selectedSubjects ?? []), subject]);
-        }
-    }
-
     return (
         <div className="flex min-h-screen items-center justify-center">
             <Card className="relative w-2xl p-5">
@@ -64,21 +40,7 @@ export default function StartPage() {
                         Choose which categories of math you want to work with:
                     </p>
                     <div className="flex flex-wrap justify-center gap-2">
-                        {Object.keys(subjectIcons).map((subject) => (
-                            <SubjectCheckbox
-                                key={subject}
-                                subject={subject as Subject}
-                                // Set checked to false on initial render, then update based on local storage
-                                checked={
-                                    isHydrated &&
-                                    (selectedSubjects?.includes(
-                                        subject as Subject,
-                                    ) ??
-                                        false)
-                                }
-                                onToggle={toggleSubject}
-                            />
-                        ))}
+                        <SubjectSelect />
                     </div>
                     Ready? Then press {'"Start Practicing"'} and get your first
                     math question.
